fix in struct_eq so that it can handle proper lists with no additional stack. Still incorrect for structures that are not proper lists

+++ utils/tinygfx.c
@@ -1022,8 +1022,8 @@ static void arc_thin_plot(image_buffer_t *img, int c_x, int c_y, int px, int py,
   if (full_circle) {
     inside = true;
   } else {
-    float cross0 = px * cap0_y - py * cap0_x;
-    float cross1 = px * cap1_y - py * cap1_x;
+    float cross0 = (float)px * cap0_y - (float)py * cap0_x;
+    float cross1 = (float)px * cap1_y - (float)py * cap1_x;
     bool inside0 = cross0 <= 0;
     bool inside1 = cross1 >= 0;
     inside = angle_is_closed ? (inside0 || inside1) : (inside0 && inside1);
@@ -1161,12 +1161,12 @@ static void arc_ring(image_buffer_t *img, int c_x, int c_y, int radius, float an
     }
 
     circle_row_draw(img, c_x, c_y, p->color, y0, x_out, x_out_r, has_gap, x_in, x_in_r,
-                    (float)angle0_cos * radius_outer, (float)angle0_sin * radius_outer,
-                    (float)angle1_cos * radius_outer, (float)angle1_sin * radius_outer,
+                    angle0_cos * (float)radius_outer, angle0_sin * (float)radius_outer,
+                    angle1_cos * (float)radius_outer, angle1_sin * (float)radius_outer,
                     angle_is_closed, filled_segment, full_circle);
     circle_row_draw(img, c_x, c_y, p->color, -y0 - 1, x_out, x_out_r, has_gap, x_in, x_in_r,
-                    (float)angle0_cos * radius_outer, (float)angle0_sin * radius_outer,
-                    (float)angle1_cos * radius_outer, (float)angle1_sin * radius_outer,
+                    angle0_cos * (float)radius_outer, angle0_sin * (float)radius_outer,
+                    angle1_cos * (float)radius_outer, angle1_sin * (float)radius_outer,
                     angle_is_closed, filled_segment, full_circle);
   }
